refactor: existing folder layout
