+++ src/pages/MainLayoutPage/MyCourseSection/index.js
@@ -1,8 +1,8 @@
 import { Container, makeStyles } from "@material-ui/core";
-import FilterBar from "../../components/MyCourseComponents/FilterBar";
-import MyCourseCard from "../../components/MyCourseComponents/MyCourseCard";
-import TagSection from "../../components/TagSection";
-import HeroSection from "../../components/HeroSection/index";
+import FilterBar from "../../../components/MyCourseComponents/FilterBar";
+import MyCourseCard from "../../../components/MyCourseComponents/MyCourseCard";
+import TagSection from "../../../components/TagSection";
+import HeroSection from "../../../components/HeroSection/index";
 function MyCourses(props) {
   const classes = useStyles();
