+++ src/pages/MainLayoutPage/WishListView/index.js
@@ -1,11 +1,11 @@
 import { Box, Container, Grid, makeStyles } from "@material-ui/core";
 import { connect } from "react-redux";
-import { wishlistDeleted } from "../../actions/wishlistActions";
-import HeroSection from "../../components/HeroSection";
-import EmptyWishlist from "../../components/WishlistComponents/EmptyWishlist";
-import WishlistFilterBar from "../../components/WishlistComponents/WishlistFilterBar";
-import WishListTagSection from "../../components/WishlistComponents/WishListTagSection";
-import MediaCard from "../../CourseMediaCard/MediaCard";
+import { wishlistDeleted } from "../../../actions/wishlistActions";
+import HeroSection from "../../../components/HeroSection";
+import EmptyWishlist from "../../../components/WishlistComponents/EmptyWishlist";
+import WishlistFilterBar from "../../../components/WishlistComponents/WishlistFilterBar";
+import WishListTagSection from "../../../components/WishlistComponents/WishListTagSection";
+import MediaCard from "../../../CourseMediaCard/MediaCard";
 
 const mapStateToProps = (state) => ({
   wishlist: state.wishlist,
@@ -40,7 +40,7 @@ const WishListView = ({ wishlist, wishlistDeleted }) => {
       <WishListTagSection />
       <Box mt={10} mb={5}>
         {wishlist.length ? (
-          <Grid container spacing={2}>
+          <Grid container spacing={4}>
             {wishlist.map((items) => {
               return (
                 <Grid key={items.id} item xs={12} sm={6} md={6} lg={4}>
